* upstream/master: (823 commits)
  Set the download link to PYPI.
  Address Issue #590 : use relative path link to about.html
  check random state in _fit_transform
  add random_state to LocallyLinearEmbedding
  MISC: fix bibtex
  Add n_jobs option to pairwise_distances and pairwise_kernels.
  euclidian_distances is to be deprecated in v0.11.
  Better default threshold for L1-regularized models.
  Remove CoefSelectTransformerMixin and use SelectorMixin instead.
  COSMIT removed unused imports
  Cosmit: remove unused imports
  Refactor in KFold.
  COSMIT pep8
  FIX: MurmurHash3 compilation on older GCC
  DOC added required versions of python, numpy and scipy to install documentation. Closes issue #579
  DOC website: added link to 0.10 docs under support.
  FIX: broken build / tests
  Move preprocessing.py to sklearn/.
  preprocessing/__init__.py -> preprocessing/preprocessing.py
  ENH: Default in Vectorizer "None" as @ogrisel suggested
  ...

+++ sklearn/feature_extraction/tests/test_image.py
@@ -54,6 +54,7 @@ def test_grid_to_graph():
     A = grid_to_graph(n_x=size, n_y=size, n_z=size, mask=mask, dtype=np.float)
     assert A.dtype == np.float
 
+
 def test_connect_regions():
     lena = sp.misc.lena()
     for thr in (50, 150):
@@ -74,11 +75,12 @@ def test_connect_regions_with_grid():
 
 
 def _downsampled_lena():
-    lena = sp.misc.lena()
+    lena = sp.misc.lena().astype(np.float32)
     lena = lena[::2, ::2] + lena[1::2, ::2] + lena[::2, 1::2] + \
            lena[1::2, 1::2]
     lena = lena[::2, ::2] + lena[1::2, ::2] + lena[::2, 1::2] + \
            lena[1::2, 1::2]
+    lena = lena.astype(np.float)
     lena /= 16.0
     return lena
